Move vscode import into settings crate
