+++ crates/settings/src/base_keymap_setting.rs
@@ -1,12 +1,9 @@
 use std::fmt::{Display, Formatter};
 
-use crate::{
-    self as settings,
-    settings_content::{BaseKeymapContent, SettingsContent},
-};
+use crate::{self as settings, settings_content::BaseKeymapContent};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use settings::{Settings, VsCodeSettings};
+use settings::Settings;
 
 /// Base key bindings scheme. Base keymaps can be overridden with user keymaps.
 ///
@@ -132,9 +129,5 @@ impl BaseKeymap {
 impl Settings for BaseKeymap {
     fn from_settings(s: &crate::settings_content::SettingsContent) -> Self {
         s.base_keymap.unwrap().into()
-    }
-
-    fn import_from_vscode(_vscode: &VsCodeSettings, current: &mut SettingsContent) {
-        current.base_keymap = Some(BaseKeymapContent::VSCode);
     }
 }
