+++ crates/zlog_settings/src/zlog_settings.rs
@@ -29,6 +29,4 @@ impl Settings for ZlogSettings {
             scopes: content.log.clone().unwrap(),
         }
     }
-
-    fn import_from_vscode(_: &settings::VsCodeSettings, _: &mut settings::SettingsContent) {}
 }
